Add face-type column to GroupEditorWidget
This commit adds an additional column (Face-Type) to the GroupEditorWidget table. This allows the user to define a conditional subset of the elements from the specified group to use for the operation.

A combo-box (Dimension of Operation) has been added to the top of the widget as it is important to specify what dimension we are working with when performing operations with specific face-types.